storybook-generator/requirements.txt
fastapi
uvicorn
openai
pillow
python-dotenv
requests
supabase
pydantic
